+++ app/repositories/task_repository.py
@@ -4,13 +4,14 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from typing import Annotated, List
 from sqlalchemy import select
 from datetime import datetime
-
+import json
 
 from app.api.schemas.user import UserInDB
 from app.db.models import Tasks
 from app.core.security import get_current_user
-from app.api.schemas.task import TaskOut, TaskCreate
-from app.core.room_manager import room_manager, redis_client
+from app.api.schemas.task import TaskOut, TaskCreate, TaskEvent
+from app.core.room_manager import redis_client
+from app.errors.task import InvalidCredentials, TaskNotFound
 
 class TaskRep(ABC):
     @abstractmethod
@@ -33,21 +34,25 @@ class SqlAlchemyTaskRep(TaskRep):
 
     async def create_new_task(self, task: TaskCreate, room: str, current_user: Annotated[UserInDB, Depends(get_current_user)]) -> TaskOut:
         user = current_user.username
-    
-        new_task = Tasks(
-            name=task.name,
-            description=task.description,
-            created=user,
-            created_at=datetime.now(),
-            status=task.status
-        )
+        try:
+            new_task = Tasks(
+                name=task.name,
+                description=task.description,
+                created=user,
+                created_at=datetime.now(),
+                status=task.status
+            )
         
-        # await redis_client.publish(room, msg=json.dumps(TaskEvent(event="task create", msg=f'{user} create new task').model_dump()))
+            self.db.add(new_task)
+            await self.db.commit()
+            await self.db.refresh(new_task)
 
-        self.db.add(new_task)
-        await self.db.commit()
-        await self.db.refresh(new_task)
-        
+        except:
+            raise InvalidCredentials(detail="Invalid credentials")
+
+        message = json.dumps(TaskEvent(event="task create", msg=f'{user} create new task').model_dump())
+        await redis_client.publish(room, message)
+
         return new_task
 
     async def get_tasks(self) -> List[TaskOut]:
@@ -55,30 +60,39 @@ class SqlAlchemyTaskRep(TaskRep):
         return tasks
 
     async def put_task(self, task_id: int, task_data: TaskCreate, room: str, current_user: Annotated[UserInDB, Depends(get_current_user)]) -> TaskOut:
+        
         cur_task = (await self.db.execute(select(Tasks).where(Tasks.id==task_id))).scalar_one_or_none()
 
+        if not cur_task:
+            raise TaskNotFound(detail="task not found")
+
         user = current_user.username
+        try:
+            cur_task.name = task_data.name
+            cur_task.description = task_data.description
+            cur_task.status = task_data.status
 
-        cur_task.name = task_data.name
-        cur_task.description = task_data.description
-        cur_task.status = task_data.status
+            await self.db.commit()
+            await self.db.refresh(cur_task)
 
-        # await redis_client.publish(room, msg=json.dumps(TaskEvent(event="modified task", msg=f'{user} modified task').model_dump()))
-
-
-        await self.db.commit()
-        await self.db.refresh(cur_task)
+        except:
+            raise InvalidCredentials(detail="Invalid credentials")
+        
+        await redis_client.publish(room, json.dumps(TaskEvent(event="modified task", msg=f'{user} modified task').model_dump()))
 
         return cur_task
 
     async def del_task(self, task_id: int, room: str, current_user: Annotated[UserInDB, Depends(get_current_user)]):
         cur_task = (await self.db.execute(select(Tasks).where(Tasks.id==task_id))).scalar_one_or_none()
 
-        user = current_user.username
+        if not cur_task:
+            raise TaskNotFound(detail="Task not found")
 
-        # await redis_client.publish(room, msg=json.dumps(TaskEvent(event="deleted task", msg=f'{user} deleted the task').model_dump()))
+        user = current_user.username
 
         await self.db.delete(cur_task)
         await self.db.commit()
         
+        await redis_client.publish(room, json.dumps(TaskEvent(event="deleted task", msg=f'{user} deleted the task').model_dump()))
+
         return {"message": "задача успешно удалена"}
